+++ STACpopulator/implementations/CMIP6_UofT/__init__.py
@@ -1,3 +1,0 @@
-from .add_CMIP6 import add_parser_args, runner
-
-__all__ = ["add_parser_args", "runner"]